Update api.php
Line 269: More than 80 characters in a line $temp was created to split character in multiple lines to resoved this issue.

Line 270:  More than 80 characters in a line $temp and $vartemp was created to split character in multiple lines to resoved this issue.

Line 271:   More than 80 characters in a line $temp and $vartemp was created to split character in multiple lines to resoved this issue.

Line 272:   More than 80 characters in a line $temp and $vartemp was created to split character in multiple lines to resoved this issue.

Line273: More than 80 characters in a line $temp and $vartemp was created to split character in multiple lines to resoved this issue.

line 274 : More than 80 characters in a line $temp and $vartemp was created to split character in multiple lines to resoved this issue.